Refactor(sunwoo/friends/#16) : 라우팅 분리
