+++ app/app.js
@@ -1,7 +1,8 @@
 "use strict";
 
 const express = require("express");
-const home = require("./src/routes/home");
+const friends = require("./src/routes/friends");
+const sign = require("./src/routes/sign");
 require("dotenv").config();
 
 const app = express();
@@ -10,6 +11,7 @@ app.use(express.static(`${__dirname}/src/public`));
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
-app.use("/", home);
+app.use("/friends", friends);
+app.use("/sign", sign);
 
 module.exports = app;
