+++ app/src/models/user/user.js
@@ -9,7 +9,7 @@ class User {
   async login(user) {
     const check = await UserStorage.login(user);
     if (check[0][0]) {
-      if (check[0][0].password === user.passWord) {
+      if (check[0][0].password === user.password) {
         return { success: true };
       } else return { success: false };
     } else return { success: false };
